requirements.txt
fastapi
uvicorn
requests
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
openai>=1.0.0